+++ validation/kubernetes/node-ttl/main.tf
@@ -6,5 +6,6 @@ provider "helm" {}
 
 module "node_ttl" {
   source                      = "../../../modules/kubernetes/node-ttl"
+  cluster_id                  = "foobar"
   status_config_map_namespace = "kube-system"
 }