+++ modules/kubernetes/spegel/main.tf
@@ -10,7 +10,7 @@ terraform {
   required_providers {
     git = {
       source  = "xenitab/git"
-      version = "0.0.2"
+      version = "0.0.3"
     }
   }
 }