+++ modules/kubernetes/node-local-dns/variables.tf
@@ -1,5 +1,9 @@
+variable "cluster_id" {
+  description = "Unique identifier of the cluster across regions and instances."
+  type        = string
+}
+
 variable "dns_ip" {
   description = "Central DNS IP"
   type        = string
-  default     = "10.0.0.10"
 }