+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [#1002](https://github.com/XenitAB/terraform-modules/pull/1002) Migrate spegel to install with Flux.
 - [#1006](https://github.com/XenitAB/terraform-modules/pull/1006) Update Git provider.
 - [#1005](https://github.com/XenitAB/terraform-modules/pull/1005) Migrate node-local-dns to install with Flux.
+- [#1007](https://github.com/XenitAB/terraform-modules/pull/1007) Migrate vpa to install with Flux.
 
 ## 2023.06.4