+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [#1006](https://github.com/XenitAB/terraform-modules/pull/1006) Update Git provider.
 - [#1005](https://github.com/XenitAB/terraform-modules/pull/1005) Migrate node-local-dns to install with Flux.
 - [#1007](https://github.com/XenitAB/terraform-modules/pull/1007) Migrate vpa to install with Flux.
+- [#1003](https://github.com/XenitAB/terraform-modules/pull/1003) Migrate gatekeeper to install with Flux.
 
 ## 2023.06.4